+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: `EdgeContextOptions.airbitzSupport`, for loading legacy Airbitz data from disk.
 - fixed: Export the `EdgeObjectTemplate` type.
 - fixed: TypeScript v5 compatibility.